Enhanced the build script to properly set the library IDs
This fix assigns proper library IDs and load paths for the dependent
libraries even if they are distributed among several directories.

This is in particular important for the stream and tool plugins
which exist in a folder next to the application but have to refer
to libraries from the Frameworks folder.

The fix consists of extending the library analysis within build4mac.py
and supplying an additional directory listing the target directories.