First solution provided for the ticket
The solution consists of separating the PluginRoot
interface from the MainWindow. As there is a
singleton getter (PluginRoot::instance) this isn't
really required and the MainWindow - PluginRoot identity
is artificial anyway: the Ruby/Python API identifies
the application with the PluginRoot. This is the better
solution as MainWindow is only available in GUI-mode
applications.