Implemented #487
The solution is to provide a new layer type: text collections.
This layer type can be manipulated by

* Boolean AND and NOT with polygons
* Polygon interact with texts
* Flatten
* Text filters

Layers of this kind are generated through "labels" now in LVS
and DRC. Text layers are more consistent than the previous
solution which consisted of pseudo-labels embedded inside
polygon layers.

Side effect: with text layers the dummy boxes for the text
markers (2x2 DBU boxes) are gone.

Text layers can participate in LVS connect and connect_global.
Text layers are the preferred way to handle texts in DRC and LVS.