Better support for point-like handles
PCells have been featuring point-like handles but
they suffered some issues because they have been
mapped to degenerated boxes:

- hardly visible
- could be destroyed using partial editing

Now, the database offers points, hence it is
possible to store points explicitly, so no
tricks need to be played to make them visible.

Editing has been implemented to some extent,
so it is possible for example to configure
handles in the properties dialogs.