Refining device extraction for smart device propagation mode
Problem in this mode is that partial devices may be
extracted. This means for example, "half" a FET many be
seen with a gate having only one source or drain shape.

In this case, we want to suppress warnings, yet ignore this
device. The upper-level extraction will take care that
nothing is lost.

In addition, the half gate should not be considered when
distributing the source/drain areas of the devices. Or it could,
but these parameters are nonsense anyway in smart device
propagation mode. We should rather skip then in that mode.