Trying to improve "grab" behavior of partial selection
Problem was: with a partial selection, clicking
in the vicinity of a non-selected edge could
make this edge the selected one. So it was
not possible to drag the selection in some
cases.

With "move mode" it was. The difference was in
the detection logic that decides where a click
is on the selection or outside.

Now, the logic is aligned, and partial mode
clicks will check whether the mouse pointer
is inside the selection bbox (plus some
margin). In that case, the click applies to
the current selection.