/ucrt64/bin added to PYTHONPATH
This addition is requred for importing
 FreeCAD, Part, and other FreeCAD submodules.

FreeCAD must be installed in the MSYS2 system before running the script "deploy-win-ucrt64.sh". This is done with command: 
 "pacman -S mingw-w64-ucrt-x86_64-freecad".

I have also done a full system upgrade using "pacman -Suy" because my MSYS2 system was very outdated and then I have built klayout 30.1. It seems to work well but there is a problem: FreCAD python modules were included in subfolders Ext and Mod of bin-release-win64-ucrt but not included in the NSIS package.